fix: 컴포넌트 리팩토링, 레이아웃 수정 등
